Enable CA1816: Dispose methods should call SuppressFinalize (#14074)
https://docs.microsoft.com/dotnet/fundamentals/code-analysis/quality-rules/ca1816

+++ test/xUnit/csharp/test_FileSystemProvider.cs
@@ -39,9 +39,18 @@ namespace PSTests.Parallel
             File.AppendAllText(testPath, testContent);
         }
 
-        void IDisposable.Dispose()
+        public void Dispose()
         {
-            File.Delete(testPath);
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }        
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                File.Delete(testPath);
+            }
         }
 
         private ExecutionContext GetExecutionContext()
